simple-website/backend/requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.26
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3